Brian's last program for RoboRams 2017
This program is for First Robotics.  The program produced a radial selector for field position to allow for the user to select and action for autonomous.  It has a gyroscope (accelerometer?) that caused the robot to stop in autonomous when it strikes something.  Teleoperated mode is fairly straightforward.  It uses a ps2 controller for the driver.